added distribute incentive function, and tested it

+++ database/migrations/2024_06_06_230256_create_distributed_incentives_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
                 ->cascadeOnDelete()
                 ->cascadeOnUpdate();
 
-            $table->float('amount');
+            $table->decimal('amount', 9, 1);
 
             $table->tinyInteger('points_amount')->unsigned();
